Establish a test suite
This includes a new test folder and pytest module. You can run pytest
from the test folder.

Includes our approximation of end-to-end tests by comparing output
FASTAs created from Stitchr example inputs. The intention is for the
baseline_*.fasta files to not be modified, and for the new_*.fasta files
to be updated after every change to the codebase when you're ready to
test.